# Simplify expression operator parsing and support multi-word operators