The operator is any letter + space + symbol combination at the end of the condition. If a condition has a complex expression, that will include a parenthesis. Even though expressions can contain spaces, there will be a ) at the end which prevents it from being pulled into the operator match.

This change also converts the operator to upper case so the generated SQL matches the expected convention.

The usual caveat applies that an identifier with spaces must be quoted through an Expression object and cannot be in a condition string.