@othercorey 2022-01-14
This could go into 4.4. I don't know of any currently working code that it would break or change.