@othercorey 2022-01-14
Technically, this will support quoting identifiers in the expression since the quote symbol will prevent it from being included in the operator.

Don't think any supported drivers have a quote symbol that's also a math symbol.