@markstory 2022-01-14
> Don't think any supported drivers have a quote symbol that's also a math symbol.

This is correct.