@othercorey 2022-01-14
Tests:
```php
    public function benchConditionParsing()
    {
        new QueryExpression(['Users.name']);
    }

    public function benchConditionOperatorParsing()
    {
        new QueryExpression(['Users.name is not']);
    }
```

Previous:
```
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| benchmark            | subject                       | set | revs | its | mem_peak | mode    | rstdev |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| QueryExpressionBench | benchConditionParsing         |     | 500  | 5   | 2.216mb  | 1.420μs | ±2.19% |
| QueryExpressionBench | benchConditionOperatorParsing |     | 500  | 5   | 2.216mb  | 1.414μs | ±1.95% |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
```

After:
```
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| benchmark            | subject                       | set | revs | its | mem_peak | mode    | rstdev |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| QueryExpressionBench | benchConditionParsing         |     | 500  | 5   | 2.182mb  | 1.427μs | ±2.56% |
| QueryExpressionBench | benchConditionOperatorParsing |     | 500  | 5   | 2.182mb  | 1.437μs | ±1.14% |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
```