@othercorey 2022-01-14
Tests:
```php
    public function benchConditionParsing()
    {
        new QueryExpression(['Users.name' => 'me']);
    }

    public function benchConditionOperatorParsing()
    {
        new QueryExpression(['Users.name is' => 'me']);
    }
```

Before:
```
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| benchmark            | subject                       | set | revs | its | mem_peak | mode    | rstdev |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| QueryExpressionBench | benchConditionParsing         |     | 500  | 5   | 2.225mb  | 3.038μs | ±0.70% |
| QueryExpressionBench | benchConditionOperatorParsing |     | 500  | 5   | 2.225mb  | 3.583μs | ±1.40% |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
```

After:
```
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| benchmark            | subject                       | set | revs | its | mem_peak | mode    | rstdev |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
| QueryExpressionBench | benchConditionParsing         |     | 500  | 5   | 2.225mb  | 3.178μs | ±0.87% |
| QueryExpressionBench | benchConditionOperatorParsing |     | 500  | 5   | 2.225mb  | 3.642μs | ±0.91% |
+----------------------+-------------------------------+-----+------+-----+----------+---------+--------+
```