@othercorey 2022-01-14
I can improve the times for expressions with no operators by checking for a space first. It adds a slight overhead to the other 2 scenarios where a single operator word is worst-case but the multi-word operator is still faster.

```
+----------------------+----------------------------------------+-----+------+-----+----------+---------+--------+
| benchmark            | subject                                | set | revs | its | mem_peak | mode    | rstdev |
+----------------------+----------------------------------------+-----+------+-----+----------+---------+--------+
| QueryExpressionBench | benchConditionParsing                  |     | 500  | 5   | 2.225mb  | 3.019μs | ±0.69% |
| QueryExpressionBench | benchConditionOperatorParsing          |     | 500  | 5   | 2.226mb  | 3.742μs | ±0.95% |
| QueryExpressionBench | benchConditionMultiWordOperatorParsing |     | 500  | 5   | 2.226mb  | 3.597μs | ±0.64% |
+----------------------+----------------------------------------+-----+------+-----+----------+---------+--------+
```