@markstory 2022-01-14
> I can improve the times for expressions with no operators by checking for a space first. 

I think optimizing for the no operator scenario is the right thing to do. Basic equality operations are more common than `is not` conditions.